gene_genie_AWS/requirements.txt
appnope==0.1.0
backcall==0.1.0
bagit==1.7.0
beautifulsoup4==4.8.0
biocode==0.9.0
biopython==1.74
CacheControl==0.11.7
certifi==2019.6.16
chardet==3.0.4
Click==7.0
coloredlogs==10.0
cwlref-runner==1.0
cwltool==1.0.20190815141648
cycler==0.10.0
decorator==4.4.0
Flask==1.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-Testing==0.7.1
future==0.17.1
humanfriendly==4.18
idna==2.8
ipython==7.7.0
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.14.1
Jinja2==2.10.1
kiwisolver==1.1.0
libgravatar==0.2.3
lockfile==0.12.2
lxml==4.4.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mypy-extensions==0.4.1
networkx==2.3
numpy==1.16.4
orderedset==2.0.1
pandas==0.25.0
parso==0.5.1
pathlib2==2.3.4
peewee==3.9.6
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
prov==1.5.1
psutil==5.6.3
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.0
python-dateutil==2.8.0
python-igraph==0.7.1.post6
pytz==2019.2
rdflib==4.2.2
rdflib-jsonld==0.4.0
requests==2.22.0
ruamel.yaml==0.15.97
scandir==1.10.0
schema-salad==4.5.20190815125611
shellescape==3.4.1
six==1.12.0
soupsieve==1.9.2
SQLAlchemy==1.3.6
taxadb==0.10.1
tqdm==4.32.2
traitlets==4.3.2
typing-extensions==3.7.4
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.5
xmltodict==0.12.0
